Enum makair_telemetry_cli::Mode

pub(crate) enum Mode {
    Debug(Debug),
    Record(Record),
    Play(Play),
    Stats(Stats),
    Control(Control),
    Storm(Storm),
    Convert(Convert),
    DisableRpiWatchdog(DisableRpiWatchdog),
}

Variants

Debug

Read telemetry from a serial port or a WebSocket server, parse it and stream result to stdout

Tuple Fields of Debug

0: Debug

Record

Read telemetry from a serial port and save bytes to a file

Tuple Fields of Record

0: Record

Play

Read telemetry from a recorded file, parse it and stream result to stdout

Tuple Fields of Play

0: Play

Stats

Read telemetry from a recorded file, parse it and compute some statistics

Tuple Fields of Stats

0: Stats

Control

Send one specific control message to a serial port, then run debug mode

Tuple Fields of Control

0: Control

Storm

Send a lot of control messages and/or bytes to a serial port

Tuple Fields of Storm

0: Storm

Convert

Read telemetry from a recorded file, parse it and convert it to another format

Tuple Fields of Convert

0: Convert

DisableRpiWatchdog

Send a control message to disable the RPi watchdog (until MCU is restarted)

Tuple Fields of DisableRpiWatchdog

0: DisableRpiWatchdog